requirements.txt
tiktoken
networkx
pyautogen
pandas
sqlalchemy
gymnasium
numpy
rich
gymnasium
pyautogen
azure-identity
azure-monitor-query
mysql-connector-python
python-Levenshtein